pyproject.toml
[tool.poetry]
name = "fishbot"
version = "0.1.0"
description = "Fishy Community Discord Bot"
authors = ["Robb Manes <robbmanes@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.1.1"
tomli = "^2.0.1"
databases = {extras = ["aiosqlite"], version = "^0.7.0"}
asyncpg = "^0.27.0"
pycord = "^0.1.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
fishbot-discord = "fishbot:main"

[build-system]
requires = [
    "poetry-core>=1.0.0",
]
build-backend = "poetry.core.masonry.api"
